Support glossary term names (OCECDR-5218)
Added a third table to the Linked Media Documents report, to show
the GlossaryTermName documents whose concept documents link to
Media documents. Also fixed the original query for showing the
concept documents themselves. That query had an unnecessary join
which was incorrectly reducing the number of rows in the report.