Make it easier to track down broken documents
The full board roster report can break if the users have marked
a board member as currently active before the document has been
validated. This change adds some logging so we'll know which
board member document we were trying to process when the report
failed (we don't get a lot of clues from the filtering module).