Fix broken audio import script (OCECDR-5361)
Intialization of the variable to collect the terms to be updated
was moved from outside the loop to inside the loop, resulting in
a condition in which only the terms and their media documents as
found in the last iteration through the loop were included for
processing. Fixed this bug and enhanced the automated test to
detect the problem in the event it should come back.